requirements.txt
aiohttp==3.8.1
sanic==22.3.1
pynacl==1.5.0
mintchoco==6.0.0
git+https://github.com/Rapptz/discord.py